Visually have Charge levels working
